added checkFirstLaunchTest in MainActivity, deleted unecessary test
