+++ services_app/src/androidTest/java/org/opendatakit/activites/LoginActivity/GeneralStateTest.java
@@ -40,22 +40,6 @@ public class GeneralStateTest extends BaseLoginActivity {
     }
 
     @Test
-    public void checkFirstStartupTest() {
-        activityScenario.onActivity(activity -> {
-            PropertiesSingleton props = CommonToolProperties.get(activity, activity.getAppName());
-            assertThat(props).isNotNull();
-
-            props.setProperties(Collections.singletonMap(CommonToolProperties.KEY_FIRST_LAUNCH, "true"));
-            activity.recreate();
-        });
-
-        onView(withId(android.R.id.button1)).inRoot(RootMatchers.isDialog()).perform(ViewActions.click());
-
-        onView(withId(R.id.inputServerUrl)).check(matches(isDisplayed()));
-        onView(withId(R.id.inputTextServerUrl)).check(matches(withText(TEST_SERVER_URL)));
-    }
-
-    @Test
     public void checkDrawerServerLoginTest() {
         onView(withId(R.id.btnDrawerOpen)).perform(ViewActions.click());
         onView(withId(R.id.drawer_server_login)).perform(ViewActions.click());
