+++ services_app/src/androidTest/java/org/opendatakit/activites/MainActivity/AnonymousStateTest.java
@@ -66,7 +66,21 @@ public class AnonymousStateTest extends BaseMainActivity {
         });
         Intents.init();
     }
+    @Test
+    public void checkFirstStartupTest() {
+        activityScenario.onActivity(activity -> {
+            PropertiesSingleton props = CommonToolProperties.get(activity, activity.getAppName());
+            assertThat(props).isNotNull();
 
+            props.setProperties(Collections.singletonMap(CommonToolProperties.KEY_FIRST_LAUNCH, "true"));
+            activity.recreate();
+        });
+
+        onView(withId(android.R.id.button1)).inRoot(RootMatchers.isDialog()).perform(ViewActions.click());
+
+        onView(withId(R.id.inputServerUrl)).check(matches(isDisplayed()));
+        onView(withId(R.id.inputTextServerUrl)).check(matches(withText(SERVER_URL)));
+    }
     @Test
     public void verifyVisibilityTest() {
         onView(withId(R.id.action_sync)).check(matches(isDisplayed()));
