+++ services_app/src/androidTest/java/org/opendatakit/activites/VerifyServerSettingsActivity/GeneralStateTest.java
@@ -56,23 +56,7 @@ public class GeneralStateTest extends BaseVerifyServerSettingActivity {
         Intents.init();
     }
 
-    @Test
-    public void checkFirstStartupTest() {
-        activityScenario.onActivity(activity -> {
-            PropertiesSingleton props = CommonToolProperties.get(activity, activity.getAppName());
-            assertThat(props).isNotNull();
-
-            props.setProperties(Collections.singletonMap(CommonToolProperties.KEY_FIRST_LAUNCH, "true"));
-            activity.recreate();
-        });
-
-        onView(withId(android.R.id.button1)).inRoot(RootMatchers.isDialog()).perform(ViewActions.click());
-
-        onView(withId(R.id.inputServerUrl)).check(matches(isDisplayed()));
-        onView(withId(R.id.inputTextServerUrl)).check(matches(withText(TEST_SERVER_URL)));
-    }
-
-    @Test
+   @Test
     public void verifyValuesTest() {
         onView(withId(R.id.tvServerUrlVerify)).check(matches(withText(TEST_SERVER_URL)));
         onView(withId(R.id.tvServerVerifyStatusVerify)).check(matches(withText(getContext().getString(R.string.not_verified))));
